Fix for handling initial INPUT value
For some reason the call to jquery ui's init function seemed to be
emptying the INPUT field, and I also couldn't see a parser anywhere for
taking data already in the INPUT field.

I've written a rudamentary parser and triggered saving the INPUT field
value before the init on datepicker is called.