Fix error when creating multiDatesPicker
I'm using ember, which extends the Array and Function properties, so they have a _super property also. This causes an error when trying to create any multiDatesPicker. To fix it, you need to iterate the dates[type] array using hasOwnProperty.

The error I get is:

Conversion from "function" format not allowed on jQuery.multiDatesPicker